#59 - install missing dep.; resolve vulnerability
The `imghdr` module was a standard module included with old versions of Python.  It is not included with Python 3.13 after being removed from a previous version.  The `standard-imghdr` module is a temporary workaround to be used until the dependency on `imghdr` can be removed.  The newer version of ArcGIS will probably resolve that.

IntelliJ IDEA alerted me that `ujson==4.1.0` had vulnerabilities/bugs, so the latest version has been installed.